remove generated file generated_add.yaml from source control
g2p/tests/public/mappings/generated_add.yaml is generated by
test_utils.py when we run the test suites, so it does not need to be in
source control. Furthermore, it's contents depend on the OS (on
Mac/Linux, it has LF linefeeds, while on Windows it has CRLF), so
running the tests on Windows shows a spurious change in the sandbox.