crg-ipa to eng-ipa mapping using merged crg-tmd and crt-dv mappings
g2p generate-mapping --ipa crg-tmd
fails because it doesn't map to crg-tmd-ipa, but rather crg-ipa.
So I cheated and temporarily appended the TMD and DV mappings into a single crg -> crg-ipa file,
used it to generate the crg-ipa -> eng-ipa mapping, add that (and only that) to the repo, and
discarded the temporary mapping.