refactor(check, tokenizer): use is_ipa() instead of endswith("-ipa")
We already have a function that tests whether a language is IPA or not.
It's still not very robust, since it still depends on the language name,
but at least it's a single place that would need to be fixed if we
improve the robustness of that definition. (See comments on PR #102
about technical debt regarding having such logic depends on language
names.)