feat: und -> und-ascii mapping calls text_unidecode.unidecode()
With this commit, we are bringing "und" back to what it was originally
intended to do, which is to take all letter characters from any alphabet
documented in the Unicode standard, and substitute it by its description
in ASCII, which is usually some approximation of how it is pronounced,
at least in some languages.

Mapping und->und-ascii is implemented as a dedicated custom function
here, not via a mapping in mappings/langs, while und-ascii->und-ipa is
done via a regular mapping.