feat: g2p generate-mapping --merge option (fixes #61)
With this commit, one can now run

    g2p generate-mapping --merge crg-tmd:crg-dv crg-ipa

to generate a crg-ipa -> eng-ipa mappings based on the union of the
crg-tmd -> crg-ipa and crg-dv -> crg-ipa mappings.

Use case: when a language has multiple writing systems, each with its own
mapping to IPA, we want the generated mapping to eng-ipa to accept the IPA
output from any of those mappings.