fix(moe): add self-mappings for k, m, n, p, s, t
This is not actually documented anywhere, but some parts of g2p assume
that putting together all the strings found on the "in" side of rules
will give you the complete inventory of graphemes for the language.
According to https://www.innu-aimun.ca/francais/grammaire/orthographe/ ,
Innu-aimun has "k", "m", "n", "p", "s", and "t" that don't show up here.
I assume they map to themselves in IPA, which is normally a good reason
to leave them out. But in g2p, it's best to include these seemingly
pointless rules.