feat: include NFC/D normalization in g2p graph
Previously, the graph generated by g2p convert was from the string
passed in after normalization. And if the path had several steps, the
graph between these steps might have been incoherent if the steps did
not all use the same normalization (see issues #158).

This commit modifies the g2p transducer class to generate the graph from
the input to convert to the output, by considering the NFC/D
normalization like a transduction step in itself, and composing the
indices it generates with the ones generated by the mapping itself.

FIXES #158