fix: make und work in g2p studio
The reason und did not work (see issue #165) is that class Transducer
relies on having type == "unidecode" to trigger the und unidecode
processing.

Solution: add an input field for "type", which defaults to "mapping",
but is set to "unidecode" when we load the und->und-ascii mapping from
cache.

Now, that input is really specific, only applicable to the und -> und-ascii
mapping, so hide it and don't let the user play with it.