fix: rules with alternations should tokenize correctly
Previously:
  "in": "é|e'"
would be considered as a whole, like that, by the tokenizer.
But the correct interpretation is to add both "é" and "e'" to the
language's input inventory.

This fixes tokenization of Mi'kmaw when í is written i' instead of í or
iꞌ (i.e., i\uA78C). All three forms should be considered equivalent, and
should get both tokenized and g2p'd the same way.