ci: replace deprecated actions/create-release by ncipollo/release-action
https://github.com/actions/create-release is deprecated so we should
stop using it. ncipollo/release-action is one of its maintained
replacements, and the one our mathieudutour/github-tag-action actions
uses in their documentation.

Linted with https://rhysd.github.io/actionlint/ but we can't really test
this without creating a new release, so we'll have to test it when it's
time to create the next release.

Fixes #200