Fix for Han/Kanji characters in Und
A recent fix to Perso-Arabic characters inadvertantly messed up Han/Kanji characters, due to an unanticipated quirk in unidecode handling of Han/Kanji characters.  Rewriting the "unidecode sanitizer" line fixes this.

Specifically, unidecode can result in a variety of non-alphabetic characters; Arabic unidecoding can result in characters like "@".  Some of these we anticipate and handle specially, others we don't.  We don't want to have an unanticipated character get downstream and result in un-interpretable errors.  So the sanitizer goes through the unidecoded string for each input character and asks "Is this alphabetic, or a special character we anticipated?"; if not, it strips it out.

This worked correctly to recognize single alphabetic characters, multiple alphabetic characters (e.g. "nu" as the output of Syllabics "ᓄ"), anticipated special characters (like "@"), and spaces.  What it didn't recognize were multi-character strings that incorporated multiple of these classes... and that's what Han/Kanji characters come out as: they're several characters and a following space.  We really needed another loop in here -- rather than checking at the STRING level whether it's alphabetic, or equal to an anticipated special character, we needed to check and strip non-alphabet, unanticipated characters at the character level.

This commit fixes this, and adds a test to test_unidecode_transducer that tests kanji inputs, which we didn't have before.