refactor: apply dhd feedback to remove dead code and unflatten the alignment
The alignment list was flattened to a list to save RAM, but now we're
only processing one at a time so that's has become pointless complexity.

The function parse_alignments is no longer used, so remove it, but take
its documentation to the functions that replaced it (and improve it).