refactor: remove explicit indices from kwk when they are monotonic
They were intended to make the mapping clearer, but in fact they hide
the result which can be seen more easily in VSCode when indices are not
there. Removing them makes no difference since they are assigned
monotonically by default.

Of course, we keep the indices when they are not monotonic.

Also, undo a few rewritten rules from the previous commit that made the
result too hard to read when the glyph combined with the curly brace
of the indices. I.e., Keep {n}\u0313{m} escaped with the \u0313 instead
of inserting its equivalent literal.