fix(str): cedilla is now the default glottal stop character
changed str-equiv to:
 - no longer change cedilla to comma
 - map the confusables to cedilla instead of comma
 - but also leave the comma as itself since it might be punctuation

changed str-to-ipa to:
 - map both cedilla and comma to glottal stop
 - document in the mapping why, and what's still not right about this.

Partially addresses #319 but does not fully fix it since we'd still want
to disambiguate instances of commas between punctuation and glottal
stop.

Fully fixing it will pose all sorts of challenges in a module without
statistical components and where tokenization assumes you can decide up
front whether a character is a letter or not. Any heuristic that maps
some but not all commas to glottal stop will likely break the ReadAlong
Studio, unless tokenization is patched at the same time to use the same
heuristics.