refactor: only declare the SCM pretend version in one place
We had forgotten to bump the heroku pretend version from 2.0 to 2.1.
This is going to happen again and again, so let's put a stop to it all
now!

The new file .SETUPTOOLS_SCM_PRETEND_VERSION has the pretend version
we'll use anywhere we have to do these fake mechanics, with the current
major.minor, and we'll make it an error in unit testing if it's out of
sync.

When we start working a the next major or minor version, we'll need that
dev tag for all this to work, which is good because this way we won't
forget! And if we try to put the new major or minor version and didn't
do it, it'll fail to run the tests.