feat: network_lite with minimal DiGraph class
This DiGraph class is meant to replace our use of networkx.
While networkx is a great library, it loads really a lot of stuff and
takes about 30MB of RAM just upon doing `import networkx`. We use maybe
a hundred lines of code from this library, so let's rewrite our own
directed graph class with just the algorithms we use.