ci: run the main tests with Python 3.9 and expand matrix testing
Licensecheck fails in 3.8 because we need the latest version and it is
only available for supported Pythons.

Keep testing with older versions via matrix-tests instead of holding the
main test suite back in the past.

And while I'm editing matrix-test, add Python 3.13 to matrix test.

Required exclusion: exclude MacOS with Python 3.7 from the matrix, since
it is no longer supported.

Matrix test: use Ubuntu 22.04 since 20.04 generates an EOL warning,
but not ubuntu-latest, because Python 3.7 is not available for U 24.04.