feat(yid): completed the Yiddish mapping in collaboration with Avery
During a working session with Avery Ozburn, Yanfei Lu and Eric Joanis, fixed
all the mapping rules for Yiddish and added a number of test words.

Known limitation: we did not try to implement voicing assimilation rules.
E.g., /gɛprʊvt/ is really pronounced /gɛprʊft/ but how voicing assimilation
happens is dialect dependent so we decided not to try to implement it.

In most dialects, a voiced consonant followed by a devoiced one gets devoiced.
In some dialects, a devoiced consonent can get voiced before a voiced one.
In some (most) dialects, the verb suffix /n/ becomes /m/ by assimilation when
it follows a /b/.

But these are issues for future improvement if this appears to be an issue
when trying to create readalongs.